requirements.txt
altgraph==0.17.4
babel==2.17.0
cabarchive==0.2.4
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
customtkinter==5.2.2
cx_Freeze==8.2.0
cx_Logging==3.2.1
darkdetect==0.8.0
filelock==3.18.0
idna==3.10
lief==0.16.5
packaging==24.2
pefile==2023.2.7
pillow==11.1.0
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
pywin32-ctypes==0.2.3
reportlab==4.4.1
requests==2.32.3
setuptools==79.0.0
striprtf==0.0.29
tkcalendar==1.6.1
urllib3==2.3.0
